// src/app/features/service-provider/service-providers/models/ServiceProvider.ts
import { ServiceProviderTypes } from '../../service-provider-types/models/ServiceProviderTypes';

export interface ServiceProvider {
  ServiceProviderId?: number; // optional because it's generated by the backend
  Name: string;
  VATNumber: string;
  CompanyRegNo: string;
  Branch: string;
  OfficeAddress: string;
  StorageAddress: string;
  TownCity: string;
  Province: string;
  ServiceProviderTypeId: number;
  DesignationNumber: string;
  Manager: string;
  AuthorisedBy: number;
  RatePerKm: number;
  RateAuthorisedOn: string; // Date as string (ISO format)
  IsActive: boolean;
  ActiveOn: string; // Date as string
  ActiveBy: number;
  IsVerified: boolean;
  VerifiedOn: string; // Date as string
  VerifiedBy: number;
  IsAccredited: boolean;
  AccreditedOn: string; // Date as string
  AccreditedBy: number;
  IsDeleted?: boolean | null; // nullable in backend
  ServiceProviderTypes?: ServiceProviderTypes | null; // nullable
  ContactDetails?: ContactDetail[] | null; // optional & nullable
}

export interface ContactDetail {
  Id?: number; // optional because it might be autogenerated
  Type: string;
  Code: string;
  Value: string;
  Name: string;
  Comments: string;
}